Add pagerank to select important insts.
Former-commit-id: 3bf8722254b0e82e742c811d3456bbda0e06e3b5

+++ MIB/src/edu/columbia/psl/cc/pojo/GraphTemplate.java
@@ -33,7 +33,7 @@ public class GraphTemplate {
 	
 	private Map<String, InstNode> writeFields;
 	
-	private int maxTime;
+	private long maxTime;
 	
 	public GraphTemplate() {
 		
@@ -166,11 +166,11 @@ public class GraphTemplate {
 		return this.threadMethodId;
 	}
 	
-	public void setMaxTime(int maxTime) {
+	public void setMaxTime(long maxTime) {
 		this.maxTime = maxTime;
 	}
 	
-	public int getMaxTime() {
+	public long getMaxTime() {
 		return this.maxTime;
 	}
